Настройка сборщика + модули
