+++ src/view/filmCardControls.js
@@ -1,0 +1,9 @@
+export const createFilmCardControls = () => ( // форма в карточке
+  `
+    <form class="film-card__controls">
+      <button class="film-card__controls-item button film-card__controls-item--add-to-watchlist">Add to watchlist</button>
+      <button class="film-card__controls-item button film-card__controls-item--mark-as-watched">Mark as watched</button>
+      <button class="film-card__controls-item button film-card__controls-item--favorite film-card__controls-item--active">Mark as favorite</button>
+    </form>
+  `
+);
