+++ src/view/filmsSection.js
@@ -1,0 +1,5 @@
+export const createFilmsSection = () => (
+  `
+      <section class="films"></section>
+  `
+);
