+++ src/view/footerStatisticsTemplate.js
@@ -1,0 +1,5 @@
+export const createFooterStatisticsTemplate = () => (
+  `
+    <p>130 291 movies inside</p>
+  `
+);
