+++ src/view/showMoreButton.js
@@ -1,0 +1,5 @@
+export const createShowMoreButton = () => ( // кнопка Show more
+  `
+    <button class="films-list__show-more">Show more</button>
+  `
+);
